Implemented alternative rendering method which does not use alpha/opacity when it is not neccessary. #define USE_ALPHA is used to alternate rendering methods. This is a neccessary step to get some kind of hardware acceleration. --------------------------------------------------------------------------------- Fixed rendering of borders of windows and selectors (working 90% correctly on non-alpha method, but it still has problems in alpha). RPG_RT actually starts blitting borders of window from position 0, 0; which is the same position as where it starts drawing the corners. This is why why it seems like it's missing a part in alpha mode. Background of windows show full 32 lines stretched. Some graphic bitmap have no transparency now. (Closes #62) Glyphs have shadow now. Scene Menu shows proper color background. --------------------------------------------------------------------------------- Moved Reader source files to a new project that generates a static lib. Player makes now use of this lib instead of compiling its own readers. --------------------------------------------------------------------------------- Moved old style static variables to a C++ anonymous namespace. Removed NULL checks before deletes. (delete will do nothing to NULL). Replaced vector<char> to vector<bool> for RAM saving.

+++ src/color.cpp
@@ -29,7 +29,7 @@ Color::Color() {
 	blue = 0;
 	alpha = 255;
 }
-Color::Color(int ired, int igreen, int iblue, int ialpha) {
+Color::Color(Uint8 ired, Uint8 igreen, Uint8 iblue, Uint8 ialpha) {
 	red = ired;
 	green = igreen;
 	blue = iblue;
@@ -60,7 +60,7 @@ bool Color::operator!=(const Color &other) const {
 ////////////////////////////////////////////////////////////
 /// Set
 ////////////////////////////////////////////////////////////
-void Color::Set(int nred, int ngreen, int nblue, int nalpha) {
+void Color::Set(Uint8 nred, Uint8 ngreen, Uint8 nblue, Uint8 nalpha) {
 	red = nred;
 	green = ngreen;
 	blue = nblue;
